feat: proper keyboard navigation
